server/pyproject.toml
[project]
name="server"
version="0.1.0"
description = "A personalized game recommendation system using vector search, graph relationships, and reinforcement learning to deliver adaptive, real-time suggestions."
readme="README.md"
requires-python = ">=3.12"
dependencies=[
    "fastapi>=0.135.3",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.2",
]
